chore: adjust jacoco thresholds
